requirements.txt
fastapi
uvicorn
tensorflow
nltk
pandas
scikit-learn
gensim
contractions
keras
tweepy
python-dotenv